refactor(util): Pull out `mod util_semver`
This `mod` is a proposal for what a new package would look like.
This needs to be split out so a future `util_manifest_schema`
package can depend on it (#12801).

This doesn't address where `RustVersion` should live (along with
`PackageIdSpec`).

This builds on the work from #12924 and #12926
